Get the testing ready for IROS paper (#2)
* Better name for pure random sampling

The goal of this function is to just record some values for comparison purposes (which we've already used).  Trying to write this in a new format

* Move domain object to simulation

* Change this for recording purposes

To allow recording of comparisons

* Define some preferences

Still need a way to manage these

* Actually do the optimization

Some stuff is hard-coded right now, and I have only tested it with random search

* Allow changing number of ei test points

* Working!

A lot of stuff printing out that I don't want, but I think this is mostly working as it should be.  Still some concern about the random seed.

* Does this do something?

I was hoping to be able to suppress the pygame output but I'm not there yet.

Co-authored-by: Michael McCourt <mccourt@sigopt.com>